server/requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-Cors==3.0.10
gunicorn==19.9.0
PyMySQL==1.0.2
Flask-SQLAlchemy==2.5.1
requests==2.26.0
Faker==9.8.3